:seedling: Switch to rspack for client bundling
Migrate from webpack to [rspack](https://rspack.rs/) for client
bundling.  This provides a significant performance improvement while
retaining the same configuration.

Update dependency versions in root and client package.json and
update the lockfile.

Resolves: #3552
