+++ client/config/rspack.dev.mts
@@ -1,25 +1,20 @@
+import { fileURLToPath } from "node:url";
 import path from "path";
 
-import ReactRefreshWebpackPlugin from "@pmmmwh/react-refresh-webpack-plugin";
-import CopyPlugin from "copy-webpack-plugin";
-import ForkTsCheckerWebpackPlugin from "fork-ts-checker-webpack-plugin";
+import { type RspackOptions, rspack } from "@rspack/core";
+import type { Configuration as DevServerConfiguration } from "@rspack/dev-server";
+import { ReactRefreshRspackPlugin } from "@rspack/plugin-react-refresh";
 import HtmlWebpackPlugin from "html-webpack-plugin";
-import ReactRefreshTypeScript from "react-refresh-typescript";
-import type { Configuration as WebpackConfiguration } from "webpack";
-import type { Configuration as DevServerConfiguration } from "webpack-dev-server";
-import { mergeWithRules } from "webpack-merge";
+import { mergeWithRules } from "rspack-merge";
+import { TsCheckerRspackPlugin } from "ts-checker-rspack-plugin";
 
-import {
-  type ClientEnv,
-  brandingAssetPath,
-  brandingStrings,
-} from "@konveyor-ui/common";
+import { type ClientEnv, brandingStrings } from "@konveyor-ui/common";
 
-import { stylePaths } from "./stylePaths";
-import commonWebpackConfiguration from "./webpack.common";
+import commonRspackConfiguration, { brandingPath } from "./rspack.common.mjs";
 
+const __dirname = fileURLToPath(new URL(".", import.meta.url));
 const pathTo = (relativePath: string) => path.resolve(__dirname, relativePath);
-const faviconPath = path.resolve(brandingAssetPath(), "favicon.ico");
+const faviconPath = path.resolve(brandingPath, "favicon.ico");
 
 /** Build the client env blob from the current process.env for dev-mode HTML injection. */
 const devClientEnv = (env: ClientEnv = process.env as unknown as ClientEnv) =>
@@ -36,7 +31,7 @@ const devClientEnv = (env: ClientEnv = process.env as unknown as ClientEnv) =>
     } as ClientEnv)
   );
 
-interface Configuration extends WebpackConfiguration {
+interface Configuration extends RspackOptions {
   devServer?: DevServerConfiguration;
 }
 
@@ -50,7 +45,7 @@ const config: Configuration = mergeWithRules({
       },
     },
   },
-})(commonWebpackConfiguration, {
+})(commonRspackConfiguration, {
   mode: "development",
   devtool: "eval-source-map",
   output: {
@@ -66,6 +61,7 @@ const config: Configuration = mergeWithRules({
     },
     hot: true,
   },
+  lazyCompilation: false,
 
   module: {
     rules: [
@@ -73,31 +69,34 @@ const config: Configuration = mergeWithRules({
         test: /\.[jt]sx?$/,
         exclude: /node_modules/,
         use: {
-          loader: "ts-loader",
+          loader: "builtin:swc-loader",
           options: {
-            transpileOnly: true, // HMR in webpack-dev-server requires transpileOnly
-            getCustomTransformers: () => ({
-              before: [ReactRefreshTypeScript()],
-            }),
+            jsc: {
+              parser: { syntax: "typescript", tsx: true },
+              transform: {
+                react: {
+                  runtime: "automatic",
+                  development: true,
+                  refresh: true,
+                },
+              },
+            },
+            detectSyntax: "auto",
           },
         },
-      },
-      {
-        test: /\.css$/,
-        include: [...stylePaths],
-        use: ["style-loader", "css-loader"],
+        type: "javascript/auto",
       },
     ],
   },
 
   plugins: [
-    new ReactRefreshWebpackPlugin(),
-    new ForkTsCheckerWebpackPlugin({
+    new ReactRefreshRspackPlugin(),
+    new TsCheckerRspackPlugin({
       typescript: {
         mode: "readonly",
       },
     }),
-    new CopyPlugin({
+    new rspack.CopyRspackPlugin({
       patterns: [
         {
           from: pathTo("../public/mockServiceWorker.js"),
@@ -129,4 +128,5 @@ const config: Configuration = mergeWithRules({
     ignored: /node_modules\/(?!@konveyor-ui\/)/,
   },
 } as Configuration);
+
 export default config;
