+++ client/config/rspack.prod.mts
@@ -1,0 +1,59 @@
+import fs from "fs";
+import { fileURLToPath } from "node:url";
+import path from "path";
+
+import { rspack } from "@rspack/core";
+import type { Configuration } from "@rspack/core";
+import HtmlWebpackPlugin from "html-webpack-plugin";
+import { merge } from "rspack-merge";
+
+import commonRspackConfiguration, { brandingPath } from "./rspack.common.mjs";
+
+const __dirname = fileURLToPath(new URL(".", import.meta.url));
+const pathTo = (relativePath: string) => path.resolve(__dirname, relativePath);
+const faviconPath = path.resolve(brandingPath, "favicon.ico");
+
+const config = merge(commonRspackConfiguration, {
+  mode: "production",
+  devtool: "nosources-source-map",
+  output: {
+    filename: "[name].[contenthash:8].min.js",
+    chunkFilename: "js/[name].[chunkhash:8].min.js",
+    assetModuleFilename: "assets/[name].[contenthash:8][ext]",
+    cssFilename: "[name].[contenthash:8].css",
+    cssChunkFilename: "css/[name].[chunkhash:8].min.css",
+  },
+
+  optimization: {
+    minimize: true,
+    minimizer: [
+      new rspack.SwcJsMinimizerRspackPlugin(),
+      new rspack.LightningCssMinimizerRspackPlugin(),
+    ],
+  },
+
+  plugins: [
+    new rspack.EnvironmentPlugin({
+      NODE_ENV: "production",
+    }),
+
+    // index.html.ejs is served at runtime by Express which injects `_env`
+    new HtmlWebpackPlugin({
+      filename: "index.html.ejs",
+      templateContent: fs.readFileSync(
+        pathTo("../public/index.html.ejs"),
+        "utf-8"
+      ),
+      favicon: faviconPath,
+      minify: {
+        collapseWhitespace: false,
+        keepClosingSlash: true,
+        minifyJS: true,
+        removeEmptyAttributes: true,
+        removeRedundantAttributes: true,
+      },
+    }),
+  ],
+} as Configuration);
+
+export default config;
