+++ client/package.json
@@ -5,12 +5,11 @@
   "license": "Apache-2.0",
   "private": true,
   "scripts": {
-    "analyze": "source-map-explorer 'dist/static/js/*.js'",
     "clean": "rimraf ./dist ./coverage",
     "prebuild": "npm run clean && npm run tsc -- --noEmit",
-    "build": "NODE_ENV=production webpack --config ./config/webpack.prod.ts",
-    "build:dev": "NODE_ENV=development webpack --config ./config/webpack.dev.ts",
-    "start:dev": "NODE_ENV=development webpack serve --config ./config/webpack.dev.ts",
+    "build": "NODE_ENV=production rspack build --config ./config/rspack.prod.mts",
+    "build:dev": "NODE_ENV=development rspack build --config ./config/rspack.dev.mts",
+    "start:dev": "NODE_ENV=development rspack dev --config ./config/rspack.dev.mts",
     "test": "NODE_ENV=test TZ=UTC jest --rootDir=. --config=./config/jest.config.ts",
     "format": "prettier --ignore-path ../.gitignore --ignore-path ../.prettierignore --ignore-unknown --write .",
     "format:check": "prettier --ignore-path ../.gitignore --ignore-path ../.prettierignore --ignore-unknown --check .",
@@ -22,7 +21,7 @@
     "@dnd-kit/core": "^6.3.1",
     "@dnd-kit/sortable": "^10.0.0",
     "@hookform/resolvers": "^2.9.11",
-    "@microlink/react-json-view": "^1.27.1",
+    "@microlink/react-json-view": "^1.31.28",
     "@monaco-editor/react": "^4.7.0",
     "@patternfly/patternfly": "6.4.0",
     "@patternfly/react-charts": "8.4.1",
@@ -33,76 +32,65 @@
     "@patternfly/react-tokens": "6.4.0",
     "@tanstack/react-query": "^4.44.0",
     "@tanstack/react-query-devtools": "^4.44.0",
-    "classnames": "^2.3.1",
-    "dayjs": "^1.11.19",
-    "fast-xml-parser": "^5.7.3",
+    "classnames": "^2.5.1",
+    "dayjs": "^1.11.23",
+    "fast-xml-parser": "^5.11.0",
     "file-saver": "^2.0.5",
     "git-url-parse": "^16.1.0",
-    "i18next": "^26.3.1",
-    "i18next-http-backend": "^4.0.0",
+    "i18next": "^26.4.0",
+    "i18next-http-backend": "^4.0.1",
     "immer": "^10.2.0",
     "js-yaml": "^4.3.1",
     "jwt-decode": "^4.0.0",
-    "lodash-es": "^4.17.21",
+    "lodash-es": "^4.18.1",
     "monaco-editor": "^0.52.2",
     "oidc-client-ts": "^3.5.0",
-    "radash": "^12.1.0",
+    "radash": "^12.1.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-error-boundary": "^6.1.2",
-    "react-hook-form": "^7.71.2",
-    "react-i18next": "^17.0.8",
+    "react-error-boundary": "^6.1.3",
+    "react-hook-form": "^7.86.0",
+    "react-i18next": "^17.0.12",
     "react-markdown": "^10.1.0",
     "react-measure": "^2.5.2",
     "react-oidc-context": "^3.3.1",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^5.3.4",
     "tinycolor2": "^1.6.0",
     "use-immer": "^0.11.0",
     "victory": "^37.3.6",
-    "xmllint-wasm": "^5.0.0",
+    "xmllint-wasm": "^5.3.0",
     "yup": "^0.32.11"
   },
   "devDependencies": {
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.6.2",
-    "@testing-library/dom": "^10.4.0",
+    "@rspack/cli": "^2.2.0",
+    "@rspack/core": "^2.2.0",
+    "@rspack/dev-server": "^2.2.1",
+    "@rspack/plugin-react-refresh": "^2.0.2",
+    "@testing-library/dom": "^10.4.1",
     "@testing-library/jest-dom": "^6.9.1",
-    "@testing-library/react": "^16.3.0",
-    "@testing-library/user-event": "^14.6.1",
+    "@testing-library/react": "^16.3.2",
+    "@testing-library/user-event": "^14.6.6",
     "@types/file-saver": "^2.0.7",
-    "@types/jest": "^29.5.4",
+    "@types/jest": "^29.5.14",
     "@types/js-yaml": "^4.0.9",
     "@types/lodash-es": "^4.17.12",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
+    "@types/react": "^18.3.31",
+    "@types/react-dom": "^18.3.7",
     "@types/react-measure": "^2.0.12",
-    "@types/react-router-dom": "^5.1.7",
+    "@types/react-router-dom": "^5.3.3",
     "@types/tinycolor2": "^1.4.6",
-    "browserslist": "^4.25.1",
-    "case-sensitive-paths-webpack-plugin": "^2.4.0",
-    "copy-webpack-plugin": "^14.0.0",
-    "css-loader": "^7.1.2",
-    "css-minimizer-webpack-plugin": "^8.0.0",
-    "fork-ts-checker-webpack-plugin": "^9.1.0",
-    "html-webpack-plugin": "^5.6.5",
+    "browserslist": "^4.28.8",
+    "html-webpack-plugin": "^5.6.8",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "jest-resolve": "^29.7.0",
     "jest-watch-typeahead": "^2.2.2",
-    "mini-css-extract-plugin": "^2.10.0",
     "monaco-editor-webpack-plugin": "^7.1.1",
     "msw": "^1.3.5",
-    "raw-loader": "^4.0.2",
-    "react-refresh": "^0.18.0",
-    "react-refresh-typescript": "^2.0.12",
-    "source-map-explorer": "^2.5.3",
-    "style-loader": "^4.0.0",
-    "terser-webpack-plugin": "^5.3.14",
-    "ts-loader": "^9.5.4",
-    "tsconfig-paths-webpack-plugin": "^4.2.0",
-    "webpack": "^5.105.3",
-    "webpack-cli": "^6.0.1",
-    "webpack-dev-server": "^5.2.6",
-    "webpack-merge": "^6.0.1"
+    "rspack-merge": "^1.0.1",
+    "ts-checker-rspack-plugin": "^1.6.1",
+    "ts-jest": "^29.4.12",
+    "ts-node": "^10.9.2"
   },
   "browserslist": {
     "production": [
