+++ package.json
@@ -48,29 +48,27 @@
     "@rollup/plugin-run": "^3.1.0",
     "@rollup/plugin-typescript": "^12.3.0",
     "@rollup/plugin-virtual": "^3.0.2",
-    "@tanstack/eslint-plugin-query": "^5.101.2",
+    "@tanstack/eslint-plugin-query": "~5.101.2",
     "@types/ejs": "^3.1.5",
-    "@types/node": "^22.19.15",
-    "concurrently": "^10.0.4",
+    "@types/node": "^22.20.1",
+    "concurrently": "^10.0.5",
     "copyfiles": "^2.4.1",
     "eslint": "^9.39.5",
     "eslint-config-prettier": "^10.1.8",
     "eslint-import-resolver-typescript": "^4.4.5",
     "eslint-plugin-import-x": "^4.17.1",
     "eslint-plugin-react": "^7.37.5",
-    "eslint-plugin-react-hooks": "^7.0.1",
-    "globals": "^16.4.0",
+    "eslint-plugin-react-hooks": "~7.0.1",
+    "globals": "^16.5.0",
     "globby": "^15.0.0",
     "husky": "^9.1.7",
     "lint-staged": "^17.3.0",
     "prettier": "^3.6.2",
     "rimraf": "^6.1.3",
-    "rollup": "^4.62.4",
+    "rollup": "^4.63.0",
     "rollup-plugin-copy": "^3.5.0",
-    "ts-jest": "^29.4.6",
-    "ts-node": "^10.9.1",
     "typescript": "~5.9.3",
-    "typescript-eslint": "^8.56.1",
+    "typescript-eslint": "^8.68.0",
     "wait-on": "^9.1.0"
   },
   "dependencies": {
@@ -83,8 +81,7 @@
     "mocha": {
       "serialize-javascript": "^7.0.6"
     },
-    "webpack-dev-server": {
-      "express": "$express"
-    }
+    "follow-redirects": "^1.16.0",
+    "qs": "^6.14.1"
   }
 }
